refactor: Moved schemas & models
