+++ src/app/api/crud/alerts.py
@@ -11,8 +11,8 @@ from sqlalchemy import and_
 import app.config as cfg
 from app.api import crud
 from app.api.routes.events import create_event
-from app.api.schemas import AlertIn, AlertOut, EventIn
 from app.db import alerts
+from app.schemas import AlertIn, AlertOut, EventIn
 
 
 async def resolve_previous_alert(device_id: int) -> Optional[AlertOut]:
