+++ src/app/api/crud/authorizations.py
@@ -8,7 +8,7 @@ from sqlalchemy import Table
 
 from app.api import crud
 from app.db import accesses
-from app.db.models import AccessType
+from app.models import AccessType
 
 
 async def is_in_same_group(table: Table, entry_id: int, group_id: int) -> bool:
