+++ src/app/api/routes/groups.py
@@ -9,9 +9,9 @@ from fastapi import APIRouter, Path, Security, status
 
 from app.api import crud
 from app.api.deps import get_current_access
-from app.api.schemas import GroupIn, GroupOut
 from app.db import groups
-from app.db.models import AccessType
+from app.models import AccessType
+from app.schemas import GroupIn, GroupOut
 
 router = APIRouter()
