+++ src/app/api/routes/users.py
@@ -10,9 +10,9 @@ from fastapi import APIRouter, Depends, Path, Security, status
 from app.api import crud
 from app.api.crud.authorizations import is_admin_access
 from app.api.deps import get_current_access, get_current_user
-from app.api.schemas import Cred, Login, UserAuth, UserCreation, UserRead
-from app.db import accesses, get_session, models, users
-from app.db.models import AccessType
+from app.db import accesses, get_session, users
+from app.models import Access, AccessType, User
+from app.schemas import Cred, Login, UserAuth, UserCreation, UserRead
 
 router = APIRouter()
 
@@ -76,12 +76,7 @@ async def fetch_users(
     if await is_admin_access(requester.id):
         return await crud.fetch_all(users)
     else:
-        retrieved_users = (
-            session.query(models.Users)
-            .join(models.Accesses)
-            .filter(models.Accesses.group_id == requester.group_id)
-            .all()
-        )
+        retrieved_users = session.query(User).join(Access).filter(Access.group_id == requester.group_id).all()
         retrieved_users = [x.__dict__ for x in retrieved_users]
         return retrieved_users
