+++ src/app/models/__init__.py
@@ -1,0 +1,10 @@
+from .access import *
+from .alert import *
+from .device import *
+from .event import *
+from .group import *
+from .installation import *
+from .media import *
+from .site import *
+from .user import *
+from .webhook import *
