+++ src/app/schemas/installations.py
